+++ flink-runtime/src/test/java/org/apache/flink/runtime/io/network/netty/ServerTransportErrorHandlingTest.java
@@ -73,13 +73,7 @@ public class ServerTransportErrorHandlingTest {
 				}
 			});
 
-		NettyProtocol protocol = new NettyProtocol() {
-			@Override
-			public ChannelHandler[] getServerChannelHandlers() {
-				return new PartitionRequestProtocol(
-					partitionManager,
-					mock(TaskEventDispatcher.class)).getServerChannelHandlers();
-			}
+		NettyProtocol protocol = new NettyProtocol(partitionManager, mock(TaskEventDispatcher.class)) {
 
 			@Override
 			public ChannelHandler[] getClientChannelHandlers() {
